Revert "[msvc] Fix warnings, mainly casting to smaller types (#270)"
This reverts commit 32ed70c95d734590b3e68cd4595a2806fd13c389.

+++ src/mbgl/renderer/layers/render_background_layer.cpp
@@ -72,7 +72,7 @@ void RenderBackgroundLayer::render(PaintParameters& parameters) {
             program.computeAllUniformValues(std::forward<decltype(uniformValues)>(uniformValues),
                                             paintAttributeData,
                                             properties,
-                                            static_cast<float>(parameters.state.getZoom()));
+                                            parameters.state.getZoom());
         const auto allAttributeBindings = program.computeAllAttributeBindings(
             *parameters.staticData.tileVertexBuffer,
             paintAttributeData,
