+++ src/mbgl/renderer/layers/render_fill_layer.cpp
@@ -103,7 +103,7 @@ void RenderFillLayer::render(PaintParameters& parameters) {
                     },
                     paintPropertyBinders,
                     evaluated,
-                    static_cast<float>(parameters.state.getZoom())
+                    parameters.state.getZoom()
                 );
                 const auto allAttributeBindings = programInstance.computeAllAttributeBindings(
                     *bucket.vertexBuffer,
@@ -196,7 +196,7 @@ void RenderFillLayer::render(PaintParameters& parameters) {
                     ),
                     paintPropertyBinders,
                     evaluated,
-                    static_cast<float>(parameters.state.getZoom())
+                    parameters.state.getZoom()
                 );
                 const auto allAttributeBindings = programInstance.computeAllAttributeBindings(
                     *bucket.vertexBuffer,
@@ -254,7 +254,7 @@ bool RenderFillLayer::queryIntersectsFeature(const GeometryCoordinates& queryGeo
             queryGeometry,
             evaluated.get<style::FillTranslate>(),
             evaluated.get<style::FillTranslateAnchor>(),
-            static_cast<float>(transformState.getBearing()),
+            transformState.getBearing(),
             pixelsToTileUnits);
 
     return util::polygonIntersectsMultiPolygon(translatedQueryGeometry.value_or(queryGeometry), feature.getGeometries());
